Refactor flashoffer demo sections
