+++ app/flashoffer-demo/src/sections/index.ts
@@ -1,0 +1,8 @@
+export * from "./types";
+export * from "./CustomizationControlsSection";
+export * from "./HeroShowcaseSection";
+export * from "./OverviewSection";
+export * from "./CtaShowcaseSection";
+export * from "./PreviewShowcaseSection";
+export * from "./FigureSpotlightSection";
+export * from "./FooterSection";
