+++ app/flashoffer-demo/src/sections/types.ts
@@ -1,0 +1,2 @@
+export type ThemePreset = "ocean" | "sunset" | "midnight";
+export type HeroAlignment = "left" | "center";
